[REF] default PGUSER and PGPASSWORD to odoo
Defaulting PGUSER to postgres does not make sense as odoo prevents
creating a database with the user `postgres` (however, there's a bug
that allows it when set by environment variable, but still it does not
work later because Odoo does not use the environment variable to find
the databases in the `list_dbs` method).

Defaulting PGPASSWORD to odoo makes sense because it's what we set in
the documentation.